+++ R/R.r
@@ -29,8 +29,7 @@ RCMD <- function(cmd, options, path = tempdir(), env_vars = NULL, ...) {
 #' @return a named character vector
 #' @export
 r_env_vars <- function() {
-  c("LC_ALL" = "C",
-    "R_LIBS" = paste(.libPaths(), collapse = .Platform$path.sep),
+  c("R_LIBS" = paste(.libPaths(), collapse = .Platform$path.sep),
     "CYGWIN" = "nodosfilewarning",
     # When R CMD check runs tests, it sets R_TESTS. When the tests
     # themeselves run R CMD xxxx, as is the case with the tests in