+++ src-i386/devtools.c
@@ -1,0 +1,6 @@
+#include <R.h>
+#include <Rdefines.h>
+
+SEXP nsreg() {
+  return R_NamespaceRegistry;
+}